chore: rename async -> write
