+++ crates/rspack_fs_node/src/lib.rs
@@ -1,7 +1,7 @@
 #![allow(clippy::unwrap_in_result)]
 
-mod r#async;
-pub use r#async::NodeFileSystem;
+mod write;
+pub use write::NodeFileSystem;
 
 mod node;
 pub use node::ThreadsafeNodeFS;
